Transient IOError (#331)
* transient IOError

* transient IOError

* move getObjectMapper back to utils

* revert unrelated

* revert unrelated

* fix warning

* revert HttpClientProperties move to http folder

* try again

* format

* disable managed streamign test

* timeout test

* t

* dis q t

* redo test

* redo test

* no retry in cloud test

* no retry in cloud test

* fix vavr not used

* stop it !

* stop it !

* stop it !

* queuedClient test again?

* no managed , queuedClienttest minus ingestFromBlob_IngestionReportMethodIsNotTable_EmptyIngestionStatus

* format

* add again ingestFromBlob_IngestionReportMethodIsNotTable_EmptyIngestionStatus

* now with retries?

* retries not vavr?

* f

* nicer

* nicer

* nicer

---------
