+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unversioned
+### Fixed
+* IOException is sometimes considered transient
 ### Added
 * Smarter way for retries with StorageAccounts
 * Support new playfab domain
+* Retries on metadata fetch
 
 ## [5.0.2] - 2023-08-24
 ### Fixed
