+++ data/src/main/java/com/microsoft/azure/kusto/data/KustoCheckedFunction.java
@@ -1,0 +1,6 @@
+package com.microsoft.azure.kusto.data;
+
+@FunctionalInterface
+public interface KustoCheckedFunction<T, R, E1 extends Throwable, E2 extends Throwable> {
+    R apply(T t) throws E1, E2;
+}
