+++ data/src/test/java/com/microsoft/azure/kusto/data/http/UncloseableStreamTest.java
@@ -1,4 +1,4 @@
-package com.microsoft.azure.kusto.data;
+package com.microsoft.azure.kusto.data.http;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
